+++ packages/airgap.js-types/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @transcend-io/airgap.js-types
 
+## 14.2.2
+
+### Patch Changes
+
+- Updated dependencies [b18f2e8]
+  - @transcend-io/privacy-types@5.1.7
+
+## 14.2.1
+
+### Patch Changes
+
+- c8e24e5: Update additionalConsentProviderId to be a string in NonTcfVendor
+
 ## 14.2.0
 
 ### Minor Changes